refactor prometheus_exporter.py and remove unnecessary code
